Rename App to hub
