+++ packages/server/src/hub/config.ts
