+++ packages/server/src/hub/handlers/handle-app-message.ts
@@ -4,7 +4,7 @@ import { ProcessProtocol } from 'magmo-wallet/lib/src/communication';
 import { AppMessage } from '../../message/app-message';
 import { startFundingProcess } from '../../wallet/db/queries/walletProcess';
 import { updateRPSChannel } from '../services/rpsChannelManager';
-
+// NOTE: This is NOT currently being used, as the hub is not going to act as a player ONLY as the hub
 export async function handleAppMessage(appMessage: AppMessage): Promise<AppMessage | undefined> {
   const { commitment: theirCommitment, signature: theirSignature } = appMessage;
   const { commitment, signature } = await updateRPSChannel(
