+++ packages/server/src/hub/handlers/handle-new-process-action.ts
@@ -3,7 +3,7 @@ import { Signature } from 'fmg-core';
 import { unreachable } from 'magmo-wallet';
 import * as communication from 'magmo-wallet/lib/src/communication';
 import { ConcludeInstigated, RelayableAction } from 'magmo-wallet/lib/src/communication';
-import { MessageRelayRequested } from '../../../../magmo-wallet-client/lib';
+import { MessageRelayRequested } from 'magmo-wallet-client';
 import { startConcludeProcess } from '../../wallet/db/queries/walletProcess';
 import { updateRPSChannel } from '../services/rpsChannelManager';
