+++ packages/server/src/hub/logging.ts
