+++ packages/server/src/test/rps_test_data.ts
@@ -17,7 +17,7 @@ import {
   RPSCommitment,
   sanitize,
   Weapon,
-} from '../app/services/rps-commitment';
+} from '../hub/services/rps-commitment';
 import {
   allocation as constAllocation,
   BEGINNING_RPS_APP_CHANNEL_NONCE,
